# [CUBRIDQA-147] improve to reduce execution time for memory leak test.