--- a/sql/_14_mysql_compatibility_2/_04_table_related/_02_alter_add_column/cases/003_many_columns.sql
+++ b/sql/_14_mysql_compatibility_2/_04_table_related/_02_alter_add_column/cases/003_many_columns.sql
@@ -13,11 +13,6 @@ insert into t (j) select j from t;
 insert into t (j) select j from t;
 insert into t (j) select j from t;
 insert into t (j) select j from t;
-insert into t (j) select j from t;
-insert into t (j) select j from t;
-insert into t (j) select j from t;
-insert into t (j) select j from t;
-insert into t (j) select j from t;
 
 select count(*) from t;