--- a/sql/_27_banana_qa/issue_12022_index_cardinality/answers/_009.answer
+++ b/sql/_27_banana_qa/issue_12022_index_cardinality/answers/_009.answer
@@ -15,14 +15,6 @@
 ===================================================
 1
 ===================================================
-1
-===================================================
-1
-===================================================
-1
-===================================================
-1
-===================================================
 0
 ===================================================
 0
@@ -32,15 +24,15 @@
 0
 ===================================================
 Table    Non_unique    Key_name    Seq_in_index    Column_name    Collation    Cardinality    Sub_part    Packed    Null    Index_type    Func    Comment    
-t     1     idx1     1     a3     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     2     b2     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     3     b3     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     4     b4     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     5     b7     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     6     b8     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     7     b10     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     8     b12     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     9     b14     A     6     null     null     YES     BTREE     null     null     
+t     1     idx1     1     a3     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     2     b2     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     3     b3     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     4     b4     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     5     b7     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     6     b8     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     7     b10     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     8     b12     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     9     b14     A     4     null     null     YES     BTREE     null     null     
 t     1     idx2     1     a2     A     3     null     null     YES     BTREE     null     null     
 t     1     idx2     2     b3     A     3     null     null     YES     BTREE     null     null     
 t     1     idx2     3     b4     A     3     null     null     YES     BTREE     null     null     
@@ -64,134 +56,116 @@ t     1     idx3     9     b10     A     6     null     null     YES     BTREE  
 t     1     idx3     10     b11     A     6     null     null     YES     BTREE     null     null     
 t     1     idx3     11     b12     A     6     null     null     YES     BTREE     null     null     
 t     1     idx3     12     b14     A     6     null     null     YES     BTREE     null     null     
-t     0     pk_t_id     1     id     A     10     null     null     NO     BTREE     null     null     
+t     0     pk_t_id     1     id     A     6     null     null     NO     BTREE     null     null     
 
 ===================================================
-10
+6
 ===================================================
-20
+12
 ===================================================
-40
+24
 ===================================================
-80
-===================================================
-160
-===================================================
-320
-===================================================
-640
-===================================================
-1280
-===================================================
-2560
-===================================================
-5120
-===================================================
-10240
-===================================================
-20480
-===================================================
-40960
+48
 ===================================================
 0
 ===================================================
 0
 ===================================================
 Table    Non_unique    Key_name    Seq_in_index    Column_name    Collation    Cardinality    Sub_part    Packed    Null    Index_type    Func    Comment    
-t     1     idx1     1     a3     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     2     b2     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     3     b3     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     4     b4     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     5     b7     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     6     b8     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     7     b10     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     8     b12     A     6     null     null     YES     BTREE     null     null     
-t     1     idx1     9     b14     A     6     null     null     YES     BTREE     null     null     
-t     1     idx2     1     a2     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     2     b3     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     3     b4     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     4     b5     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     5     b6     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     6     b7     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     7     b8     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     8     b9     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     9     b10     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     10     b11     A     42     null     null     YES     BTREE     null     null     
-t     1     idx2     11     b12     A     42     null     null     YES     BTREE     null     null     
-t     1     idx3     1     a1     A     84     null     null     NO     BTREE     null     null     
-t     1     idx3     2     b3     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     3     b4     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     4     b5     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     5     b6     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     6     b7     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     7     b8     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     8     b9     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     9     b10     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     10     b11     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     11     b12     A     84     null     null     YES     BTREE     null     null     
-t     1     idx3     12     b14     A     84     null     null     YES     BTREE     null     null     
-t     1     idx4     1     a1     A     140     null     null     NO     BTREE     null     null     
-t     1     idx4     2     a2     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     3     b1     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     4     b2     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     5     b3     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     6     b4     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     7     b5     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     8     b6     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     9     b7     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     10     b8     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     11     b10     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     12     b11     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     13     b12     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     14     b13     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     15     null     A     140     null     null     YES     BTREE      trim(both  from [t].[b14])     null     
-t     1     idx4     16     b9     A     140     null     null     YES     BTREE     null     null     
-t     1     idx4     17     a3     A     140     null     null     YES     BTREE     null     null     
-t     0     pk_t_id     1     id     A     81920     null     null     NO     BTREE     null     null     
+t     1     idx1     1     a3     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     2     b2     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     3     b3     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     4     b4     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     5     b7     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     6     b8     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     7     b10     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     8     b12     A     4     null     null     YES     BTREE     null     null     
+t     1     idx1     9     b14     A     4     null     null     YES     BTREE     null     null     
+t     1     idx2     1     a2     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     2     b3     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     3     b4     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     4     b5     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     5     b6     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     6     b7     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     7     b8     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     8     b9     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     9     b10     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     10     b11     A     15     null     null     YES     BTREE     null     null     
+t     1     idx2     11     b12     A     15     null     null     YES     BTREE     null     null     
+t     1     idx3     1     a1     A     30     null     null     NO     BTREE     null     null     
+t     1     idx3     2     b3     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     3     b4     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     4     b5     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     5     b6     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     6     b7     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     7     b8     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     8     b9     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     9     b10     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     10     b11     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     11     b12     A     30     null     null     YES     BTREE     null     null     
+t     1     idx3     12     b14     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     1     a1     A     30     null     null     NO     BTREE     null     null     
+t     1     idx4     2     a2     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     3     b1     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     4     b2     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     5     b3     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     6     b4     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     7     b5     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     8     b6     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     9     b7     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     10     b8     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     11     b10     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     12     b11     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     13     b12     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     14     b13     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     15     null     A     30     null     null     YES     BTREE      trim(both  from [t].[b14])     null     
+t     1     idx4     16     b9     A     30     null     null     YES     BTREE     null     null     
+t     1     idx4     17     a3     A     30     null     null     YES     BTREE     null     null     
+t     0     pk_t_id     1     id     A     96     null     null     NO     BTREE     null     null     
 
 ===================================================
 index_cardinality('t', 'idx1', 0)    
-6     
+4     
 
 ===================================================
 index_cardinality('t', 'idx2', 1)    
-42     
+15     
 
 ===================================================
 index_cardinality('t', 'idx3', 2)    
-84     
+30     
 
 ===================================================
 index_cardinality('t', 'idx1', 3)    
-6     
+4     
 
 ===================================================
 index_cardinality('t', 'idx2', 4)    
-42     
+15     
 
 ===================================================
 index_cardinality('t', 'idx3', 5)    
-84     
+30     
 
 ===================================================
 index_cardinality('t', 'idx4', 6)    
-140     
+30     
 
 ===================================================
 index_cardinality('t', 'idx1', 7)    
-6     
+4     
 
 ===================================================
 index_cardinality('t', 'idx2', 8)    
-42     
+15     
 
 ===================================================
 index_cardinality('t', 'idx3', 9)    
-84     
+30     
 
 ===================================================
 index_cardinality('t', 'idx4', 10)    
-140     
+30     
 
 ===================================================
 index_cardinality('t', 'idx1', 11)    
@@ -207,15 +181,15 @@ null
 
 ===================================================
 index_cardinality('t', 'idx4', 14)    
-140     
+30     
 
 ===================================================
 index_cardinality('t', 'idx4', 15)    
-140     
+30     
 
 ===================================================
 index_cardinality('t', 'idx4', 16)    
-140     
+30     
 
 ===================================================
 index_cardinality('t', 'idx4', 17)