… sql/_10_connect_by/_11_hq_hash_partition/answers/_002_performance_partition.answer
… sql/_11_codecoverage/answers/outerjoin.answer
… sql/_13_issues/_13_2h/answers/bug_bts_11504.answer
… sql/_13_issues/_13_2h/cases/bug_bts_11504.sql
… sql/_14_mysql_compatibility_2/_04_table_related/_02_alter_add_column/answers/003_many_columns.answer
… sql/_23_apricot_qa/_02_performance/_03_index_skip_scan/answers/function_index_skip_scan_700000_with_null.answer
… sql/_23_apricot_qa/_02_performance/_05_partition_enhancement/answers/_22_insert_range_01.answer
… sql/_24_aprium_qa/_03_other/issue_7441_MRO/answers/_09_mro_test.answer
… sql/_24_aprium_qa/_03_other/issue_7441_MRO/cases/_09_mro_test.sql
… sql/_24_aprium_qa/_03_other/issue_8337_insertDuplicatePerformance/answers/_13_odku_stress.answer
… sql/_27_banana_qa/issue_12022_index_cardinality/cases/_009.sql
… sql/_28_features_930/issue_7508_iss_groupby/answers/_01_function.answer
… sql/_28_features_930/issue_7508_iss_groupby/answers/_02_function_buildvalue.answer