We improved test cases on the execution time as below for memory leak test. We also confirmed each query plan that they keep the same as a result.

test case | execution time (before) | execution time (after)
------------ | ------------- | -------------
sql/_28_features_930/issue_7508_iss_groupby/cases/_01_function.sql |  67 mins | 1 min
sql/_24_aprium_qa/_03_other/issue_7441_MRO/cases/_09_mro_test.sql | 67 mins | 3 mins
sql/_23_apricot_qa/_02_performance/_05_partition_enhancement/cases/_22_insert_range_01.sql | 62 mins| 3 mins
sql/_14_mysql_compatibility_2/_04_table_related/_02_alter_add_column/cases/003_many_columns.sql | 59 mins | 1.8 mins|
sql/_28_features_930/issue_7508_iss_groupby/cases/_02_function_buildvalue.sql | 56 mins | 0.7 min|
sql/_13_issues/_13_2h/cases/bug_bts_11504.sql | 50 mins | 1.8 mins|
sql/_10_connect_by/_11_hq_hash_partition/cases/_002_performance_partition.sql | 82 mins| 10 mins
sql/_11_codecoverage/cases/outerjoin.sql | 602 mins | 132 secs
sql/_24_aprium_qa/_03_other/issue_8337_insertDuplicatePerformance/cases/_13_odku_stress.sql | 235 mins | 159 secs
sql/_27_banana_qa/issue_12022_index_cardinality/cases/_009.sql | 221 mins | 27 secs
sql/_23_apricot_qa/_02_performance/_03_index_skip_scan/cases/function_index_skip_scan_700000_with_null.sql | 219 mins | 76 secs